add link to contracts timesheet

+++ app/controllers/employer/contracts_controller.rb
@@ -10,10 +10,17 @@ class Employer::ContractsController < ApplicationController
     @contracts = current_user.employer_profile.contracts.employer_visible.order(created_at: :desc)
 
     redirect_to root_path if @contracts.hire_group.blank?
-    return unless session[:review_by_contract].present?
+    return unless session[:review_by_contract].present? || session[:timesheet_id].present?
 
-    @contract = @contracts.find { |c| c.id == session[:review_by_contract].to_i }
-    session.delete(:review_by_contract)
+    if session[:timesheet_id].present?
+      timesheets = Timesheet.related_to_contracts(current_user.employer_profile.contracts.employer_visible.ids)
+      @timesheet = timesheets.find_by(id: session[:timesheet_id])
+      @contract = @timesheet.contract
+      session.delete(:timesheet_id)
+    else
+      @contract = @contracts.find { |c| c.id == session[:review_by_contract].to_i }
+      session.delete(:review_by_contract)
+    end
   end
 
   def withdraw_offer
